feat(app): move sync to chain logic to its own component
